Moved the OnUpdatedDataStr to be declared, as it was subscribed each time OnAfterRender was called, which happends on the Updatechart call.
Renamed DataSetsStr to DataStr due to including all of data, which included the labels - to properly reset when clicking "I messed up".

Moved DataDisplay and ElementRefExtensions to a shared library as it is used in varius samples.

Added a serverside version which has the same pages as the client version - only because its easier to debug C# code serverside.
I couldn't figure out a way to only define the pages one place.

Javascript files are copied to the server project aswell, as it cannot reference static embedded files yet, from libraries.

Bug: Instance js interop is not working for some reason.