Add tests, special case for int deserialization, comments
When deserializing "0" or "123" you would get an int. If the enum has a constructor for double but not for int, it will now convert the value to a double. This might be a rather common case so this special check is worth it.
Also add lots of comments and tests plus a very minor performance enhancement. XML-docs updated.