Remove ArgValidation-helper
This class had mostly helper methods for the now replaced click and hover handlers. The null or whitespace checks are inlined everywhere so this class isn't used anymore. This was also mentioned in this [PR-comment](https://github.com/mariusmuntean/ChartJs.Blazor/pull/94#issuecomment-605660344). If we have a helper class for argument validation, we need to actually use it.