Stringify before the interop-call
This commit removed cycle.js and uses a replacer method along with JSON.stringify to serialize the object already before actually passing it to C#. This is a performance improvement because blazor doesn't have to deserialize the object, only that string. We want a string anyway since we deserialize using json.net (custom converters, consistency, etc) so we can get rid of that conversion. If you don't have a type to deserialize the callback parameters to, you now use json.net objects like JObject, JArray and JValue instead of JsonElement.