Replace data-array-reference instead of updating it
Chart.js records what changes where made to the data and animates the chart accordingly. To get the same behaviour we would also have to record, emit and apply such metadata which can be considered as a later feature. For now, we just replace the reference so Chart.js doesn't think all the data is fresh.