Make everything nullable, remove hardcoded default values
Chart.js will chose the correct default and even respect user-defined defaults when the property is not specified. Therefore we shouldn't serialize anything unless there's a non-default value assigned to it (aka make the properties nullable). This is also a performance improvement since less data has to be serialized and transferred but that gain is tiny.
This is of course a breaking change but shouldn't affect much user code.